refactor(api): move praise/utils.ts into praise/utils/core.ts
